feat(backend): simplify fact package
